requirement.txt
simplejson
uwsgi
CloudFlare
django-cors-headers
APScheduler